+++ packages/ui/src/index.ts
@@ -34,3 +34,4 @@ export * from './components/ConfirmDialog';
 export * from './components/StatusDot';
 export * from './components/ThemeSurface';
 export * from './components/Table';
+export * from './components/DateRangePicker';